Add Built.generateCopyMethod attribute
Add a generateCopyMethod attribute to the Built that allows classes
to "opt in" to having a static copy method added to the generated
builder class. If this attribute is set to true, then a static method
will be generated in the form:

public static FooBuilder fromFoo(Foo foo);

By default, this attribute is false.